Potential fix for #75 and #129.
Function OpenAIApiAxiosParamCreator now has a default parameter that instantiates a new Configuration.

I removed the use of creating new instances of FormData() since configuration is now guaranteed. I also removed any "if statements" that were testing if the configuration parameter was supplied.

This ensures that the use of the .getHeaders() method will not return TypeErrors due to FormData() instances not having that method.

Upon testing the scenarios from the bug reports in #75 and #129 I was able to prevent the TypeError.
Upon testing the scenario of not providing credentials when creating a new Configuration, the proper error was still returned.